+++ CHANGELOG.md
@@ -5,11 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Support for serializing a `Toml` object to a valid TOML string
+
 ### Changed
 - All accessor methods now return `Optionals` and do not throw.  Notably, this
   means that `TomlError.KeyError` is no longer thrown if a missing key is
   requested.
 - `Toml.keys` now returns a `Set<[String]> of value key names
+- `Toml.description` now returns a valid TOML string
 - Moved `Toml.setValue(_:[String])` from  the private API to the public API
 
 ## [0.3.0] - 2016-08-19